refactor: OrderTableValidator 인터페이스 및 구현체 생성으로 의존성 분리
