+++ src/main/java/kitchenpos/order/application/OrderTableValidator.java
@@ -1,0 +1,7 @@
+package kitchenpos.order.application;
+
+import kitchenpos.order.dto.OrderCreateRequest;
+
+public interface OrderTableValidator {
+    Long validOrderTableAndGet(final OrderCreateRequest orderCreateRequest);
+}
